editoast: client: move module `roles` in `authorization`
