+++ editoast/src/client/mod.rs
@@ -7,7 +7,6 @@ pub mod import_rolling_stock;
 pub mod infra_commands;
 mod openfga_config;
 mod postgres_config;
-pub mod roles;
 pub mod runserver;
 pub mod search_commands;
 pub mod stdcm_search_env_commands;
@@ -17,6 +16,7 @@ pub mod user;
 mod valkey_config;
 
 use authorization::grants::GrantsCommand;
+use authorization::roles::RolesCommand;
 use clap::Parser;
 use clap::Subcommand;
 use clap::ValueEnum;
@@ -25,7 +25,6 @@ use import_rolling_stock::ImportRollingStockArgs;
 use infra_commands::InfraCommands;
 use openfga_config::OpenfgaConfig;
 pub use postgres_config::PostgresConfig;
-use roles::RolesCommand;
 use runserver::CoreArgs;
 use runserver::RunserverArgs;
 use search_commands::SearchCommands;
