+++ editoast/src/main.rs
@@ -20,6 +20,8 @@ use client::Color;
 use client::Commands;
 use client::authorization::grants;
 use client::authorization::grants::GrantsCommand;
+use client::authorization::roles;
+use client::authorization::roles::RolesCommand;
 use client::electrical_profiles_commands::*;
 use client::garbage_collector;
 use client::group;
@@ -28,8 +30,6 @@ use client::healthcheck::healthcheck_cmd;
 use client::import_rolling_stock::*;
 use client::infra_commands::*;
 use client::print_openapi;
-use client::roles;
-use client::roles::RolesCommand;
 use client::runserver::runserver;
 use client::search_commands::*;
 use client::stdcm_search_env_commands::handle_stdcm_search_env_command;
